requirements.txt
todoist-python >= 8.1.1
PyYAML >= 3.12
